+++ Dev/Dev2.Runtime.Services/ServiceModel/Data/Workflow.cs
@@ -1,7 +1,7 @@
 #pragma warning disable
 /*
 *  Warewolf - Once bitten, there's no going back
-*  Copyright 2020 by Warewolf Ltd <alpha@warewolf.io>
+*  Copyright 2021 by Warewolf Ltd <alpha@warewolf.io>
 *  Licensed under GNU Affero General Public License 3.0 or later.
 *  Some rights reserved.
 *  Visit our website for more information <http://warewolf.io/>
@@ -139,7 +139,7 @@ namespace Dev2.Runtime.ServiceModel.Data
                     wfTree.Add(WorkflowNodeFrom(activityTrue));
                 }
 
-                if (!IsFlowStep(node.True))
+                if (!IsFlowStep(node.True) && node.True != null)
                 {
                     wfTree.Add(GetWorkflowNodeFrom(node.True));
                 }
@@ -150,7 +150,7 @@ namespace Dev2.Runtime.ServiceModel.Data
                     wfTree.Add(WorkflowNodeFrom(activityFalse));
                 }
 
-                if (!IsFlowStep(node.False))
+                if (!IsFlowStep(node.False) && node.False != null)
                 {
                     wfTree.Add(GetWorkflowNodeFrom(node.False));
                 }